Refactor user state store builder
